enabled LensingLayer in SkyLens; improvements in SourceCatalog

+++ include/Observation.h
@@ -28,7 +28,7 @@ namespace skylens {
     /// the pixel positions are taken from the Grid of \p im, such that one 
     /// can specify the particular field of the observation. This requires
     /// a proper setup of WCS of \p im.
-    void makeImage(shapelens::Image<double>& im, bool auto_adjust = true) const;
+    void makeImage(shapelens::Image<float>& im, bool auto_adjust = true) const;
     /// Get total transmission \f$T(\lambda)\f$, including atmospheric
     /// extinction.
     /// According to // according to Grazian et al. (2004), eq. 3
@@ -47,7 +47,7 @@ namespace skylens {
     bool hasNoise;
     filter total_air;
     double time, ron, flat_field;
-    void addNoise(const gsl_rng* r, double& signal) const;
+    void addNoise(const gsl_rng* r, float& signal) const;
     
   };
 } // end namespace
